+++ vector/src/main/java/im/vector/VectorApp.java
@@ -405,8 +405,9 @@ public class VectorApp extends MultiDexApplication {
         for (MXSession session : sessions) {
             if (session.isAlive()) {
                 session.setIsOnline(false);
-                session.setSyncDelay(pushManager.isBackgroundSyncAllowed() ? pushManager.getBackgroundSyncDelay() : 0);
-                session.setSyncTimeout(pushManager.getBackgroundSyncTimeOut());
+                //This is manage by the event stream service
+//                session.setSyncDelay(pushManager.isBackgroundSyncAllowed() ? pushManager.getBackgroundSyncDelay() : 0);
+//                session.setSyncTimeout(pushManager.getBackgroundSyncTimeOut());
 
                 // remove older medias
                 if ((System.currentTimeMillis() - mLastMediasCheck) < (24 * 60 * 60 * 1000)) {
@@ -523,8 +524,9 @@ public class VectorApp extends MultiDexApplication {
             for (MXSession session : sessions) {
                 session.getMyUser().refreshUserInfos(null);
                 session.setIsOnline(true);
-                session.setSyncDelay(0);
-                session.setSyncTimeout(0);
+                //Let the eventStream service manage that
+//                session.setSyncDelay(0);
+//                session.setSyncTimeout(0);
                 addSyncingSession(session);
             }